move class Password from postgresqldb.py to tpch_pgsql.py
